Fixed multiple IOconfOutput rows of the same type overriding the BoardSettings. IOconfInput and IOconfOutput now uses the same GetMap method (with the correct behavior).
